refactor(cli): limit explicit prompt cache breakpoint changes to Vercel AI SDK path
